requirements.txt
web3==6.11.0
eth-account==0.9.0
redis==5.0.8
aiohttp==3.12.15
websockets==15.0.1
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
python-telegram-bot==20.5
discord-webhook==1.3.0
scikit-learn==1.3.0
flask==2.3.3
gunicorn==21.2.0
psutil==5.9.5
cryptography==41.0.7
pycryptodome==3.18.0
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
watchdog==3.0.0
fastapi==0.116.1
uvicorn==0.23.2
httpx==0.24.1
pendulum==2.1.2
# Added missing dependencies
prometheus-client==0.17.1
slowapi==0.1.8
eth-abi==4.2.1
PyJWT==2.9.0
pip-audit==2.7.0
bandit==1.7.9
ruff==0.6.9
